moved sql related functions to a separate file
